refactor: перенес Обработчики в директорию internal
